Fix: change for loop to while loop
this change fixes the issue of the user entering a falsy input
but still using a game round.
now instead of using up a game round, the user will be asked
to enter a valid input until one is given.
this ensures that every round will always have valid inputs
and always returns a valid result of win lost or draw